requirements.txt
fastapi
uvicorn[standard]
passporteye
python-dateutil
python-multipart
